Add initial implementation of the library
It is still untested, likely buggy and slow. All the algorithms
were implemented in the "simple" way so that now some tests can
be developed and after that more thought can be given to performance.

It still contains a lot of (likely premature) optimizations for
different platforms.

Any progress is now blocked by adding Unit Tests, Performance Tests
and some form of CI.